refactoring: extract the code related to links to a new directory "links"
