+++ src/api/consumer/lib/link.js
@@ -1,7 +1,8 @@
 // @flow
 import { loadConsumer, Consumer } from '../../../consumer';
+import { linkAllToNodeModules } from '../../../links';
 
 export default (async function linkAction() {
   const consumer: Consumer = await loadConsumer();
-  return consumer.linkAll();
+  return linkAllToNodeModules(consumer);
 });
