+++ src/links/index.js
@@ -1,0 +1,17 @@
+import linkComponentsToNodeModules from './node-modules-linker';
+import { linkAllToNodeModules, writeLinksInDist, reLinkDirectlyImportedDependencies } from './linker';
+import {
+  writeEntryPointsForComponent,
+  writeDependencyLinks,
+  generateEntryPointDataForPackages
+} from './link-generator';
+
+export {
+  linkComponentsToNodeModules,
+  linkAllToNodeModules,
+  writeLinksInDist,
+  reLinkDirectlyImportedDependencies,
+  writeEntryPointsForComponent,
+  writeDependencyLinks,
+  generateEntryPointDataForPackages
+};
